+++ functions/src/services/auth.service.ts
